✅ Update tests to handle coverage
